putting wishlist edit page modal into the same file so that the custom modal can be renamed to useModal
